Rename module `metadata` to `_metadata`
All the other modules are private, so there is no reason why this
shouldn't be either.
