+++ src/frequenz/client/microgrid/_client.py
@@ -48,8 +48,8 @@ from ._component_data import (
 )
 from ._connection import Connection
 from ._constants import RECEIVER_MAX_SIZE
+from ._metadata import Location, Metadata
 from ._retry import LinearBackoff, RetryStrategy
-from .metadata import Location, Metadata
 
 DEFAULT_GRPC_CALL_TIMEOUT = 60.0
 """The default timeout for gRPC calls made by this client (in seconds)."""
