+++ src/frequenz/client/microgrid/_metadata.py
